full admin pages

+++ drizzle/0002_condemned_gwen_stacy.sql
@@ -1,1 +1,9 @@
-ALTER TABLE "personnel" DROP COLUMN "tg_username";
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT 1 FROM information_schema.columns 
+        WHERE table_schema = 'public' AND table_name = 'personnel' AND column_name = 'tg_username'
+    ) THEN
+        ALTER TABLE "personnel" DROP COLUMN "tg_username";
+    END IF;
+END $$;
